sysdump even if kubearmor pods are not found
currently, the sysdump expects the kubearmor daemon + pods to be
mandatorily present in the k8s. If not present, the sysdump errors out
and no zip file is produced. karmor sysdump could also be used in
cases where the user might just want to provide the snapshot of current
k8s cluster on which they intend to deploy kubearmor.
Similarly, sysdump is used in the context where kubearmor might be used
in host process mode (for e.g, dev env).

Signed-off-by: Rahul Jadhav <nyrahul@gmail.com>